linenoise core extracted
