+++ Makefile
@@ -5,7 +5,7 @@ all : linenoise_example
 
 linenoise.o : linenoise.h
 
-liblinenoise.a : linenoise.o lnTermPosix.o lnHist.o
+liblinenoise.a : linenoise.o lnCore.o lnTermPosix.o lnHist.o
 	$(AR) rcs $@ $^
 
 linenoise_example: example.c liblinenoise.a
