+++ linenoise.h
@@ -37,10 +37,33 @@
 #ifndef __LINENOISE_H
 #define __LINENOISE_H
 
+#include <sys/types.h>
+
+/* The linenoiseState structure represents the state during line editing.
+ * We pass this state to functions implementing specific editing
+ * functionalities. */
+struct linenoiseState {
+    struct lnTerminal *lnTerm;
+    struct lnHistory *lnHist;
+    char *buf;          /* Edited line buffer. */
+    size_t buflen;      /* Edited line buffer size. */
+    const char *prompt; /* Prompt to display. */
+    size_t plen;        /* Prompt length. */
+    size_t pos;         /* Current cursor position. */
+    size_t oldpos;      /* Previous refresh cursor position. */
+    size_t len;         /* Current edited line length. */
+    size_t cols;        /* Number of columns in terminal. */
+    size_t maxrows;     /* Maximum num of rows used so far (multiline mode) */
+    size_t history_index;
+};
+
 typedef struct linenoiseCompletions {
   size_t len;
   char **cvec;
 } linenoiseCompletions;
+
+int linenoiseEdit(struct lnTerminal *lnTerm, struct lnHistory *lnHist, 
+	char *buf, size_t buflen, const char *prompt);
 
 typedef void(linenoiseCompletionCallback)(const char *, linenoiseCompletions *);
 void linenoiseSetCompletionCallback(linenoiseCompletionCallback *);
